Fixing page object file location and setting demo dummy test to be ignored
